plot core

+++ rename_pl.py
@@ -8,8 +8,8 @@ for dir in os.listdir(TOPO):
     for i in ['a0.1', 'a0.2', 'a0.3', 'a0.4', 'a0.5', 'a0.6', 'a0.7', 'a0.8', 'a0.9']:
         samp = os.path.join(TOPO ,os.path.join(dir,f'graph/{i}_shortest_paths.json'))
         count = os.path.join(TOPO ,os.path.join(dir, f'graph/{i}_path_counts.pkl'))
-        os.remove(samp)
-        os.remove(count)
+        if os.path.exists(samp):os.remove(samp)
+        if os.path.exists(count):os.remove(count)
 
         '''
     if os.path.exists(pl):
